p3 seemingly done

+++ ex01/p3/calc_primo.c
@@ -1,7 +1,7 @@
-int calc_primo(int n) {
+int calc_primo(long int n) {
   int i = 0;
 
-  for(i = 2; i < n; ++i)
+  for (i = 2; i < n; ++i)
     if (n % i == 0)
       return 0;
